# docs/integraciones/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chamán Integraciones",
    "version": "1.0.0-pilot",
    "description": "Contrato multicliente. Piloto activo en Testing; preparación de Producción pendiente de activación y validación de paridad fenológica. Sin acceso a base de datos ni fórmulas. Ver README/OPERACION."
  },
  "servers": [
    {
      "url": "https://testing-api-testing.up.railway.app/sdc-quimica-test/integraciones/v1",
      "description": "Testing: confirmar activación y prefijo antes de usar"
    },
    {
      "url": "https://chaman-api-production.up.railway.app/sdc-quimica/integraciones/v1",
      "description": "Producción: no habilitada para el cliente hasta completar y autorizar la activación"
    }
  ],
  "security": [
    {
      "IntegrationKey": []
    }
  ],
  "paths": {
    "/servicios": {
      "get": {
        "operationId": "consultarServicios",
        "summary": "Permisos y capacidades de la integración",
        "responses": {
          "200": {
            "description": "Configuración pública",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Identificador para soporte"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Servicios"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida/vencida",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Servicio/operador/plan no habilitado",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no disponible",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto; ver instructivo",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de consultas",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "60 segundos"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible; reintentar sin cambiar IDs",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/catalogos/semillas": {
      "get": {
        "operationId": "consultarSemillas",
        "summary": "Semillas; requiere catalogos:leer",
        "parameters": [
          {
            "in": "query",
            "name": "cultivo",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pagina",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 10000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de hasta 100 semillas",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Identificador para soporte"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalogo"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida/vencida",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Servicio/operador/plan no habilitado",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no disponible",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto; ver instructivo",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de consultas",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "60 segundos"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible; reintentar sin cambiar IDs",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/productores/{idExterno}": {
      "parameters": [
        {
          "in": "path",
          "name": "idExterno",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "minLength": 1,
            "maxLength": 80
          }
        }
      ],
      "get": {
        "operationId": "consultar_productores",
        "summary": "Consulta de productores; requiere estructura:leer",
        "responses": {
          "200": {
            "description": "Recurso propio",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Identificador para soporte"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recurso"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida/vencida",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Servicio/operador/plan no habilitado",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no disponible",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto; ver instructivo",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de consultas",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "60 segundos"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible; reintentar sin cambiar IDs",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "registrar_productores",
        "summary": "Alta repetible de productores; requiere estructura:crear",
        "description": "No es actualización: otro contenido bajo el mismo ID devuelve 409. No crear IDs nuevos para reintentar un timeout.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductorInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alta confirmada o repetida, indicado por creado",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Identificador para soporte"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recurso"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida/vencida",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Servicio/operador/plan no habilitado",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no disponible",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto; ver instructivo",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de consultas",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "60 segundos"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible; reintentar sin cambiar IDs",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/establecimientos/{idExterno}": {
      "parameters": [
        {
          "in": "path",
          "name": "idExterno",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "minLength": 1,
            "maxLength": 80
          }
        }
      ],
      "get": {
        "operationId": "consultar_establecimientos",
        "summary": "Consulta de establecimientos; requiere estructura:leer",
        "responses": {
          "200": {
            "description": "Recurso propio",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Identificador para soporte"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recurso"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida/vencida",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Servicio/operador/plan no habilitado",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no disponible",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto; ver instructivo",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de consultas",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "60 segundos"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible; reintentar sin cambiar IDs",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "registrar_establecimientos",
        "summary": "Alta repetible de establecimientos; requiere estructura:crear",
        "description": "No es actualización: otro contenido bajo el mismo ID devuelve 409. No crear IDs nuevos para reintentar un timeout.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstablecimientoInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alta confirmada o repetida, indicado por creado",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Identificador para soporte"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recurso"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida/vencida",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Servicio/operador/plan no habilitado",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no disponible",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto; ver instructivo",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de consultas",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "60 segundos"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible; reintentar sin cambiar IDs",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/lotes/{idExterno}": {
      "parameters": [
        {
          "in": "path",
          "name": "idExterno",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "minLength": 1,
            "maxLength": 80
          }
        }
      ],
      "get": {
        "operationId": "consultar_lotes",
        "summary": "Consulta de lotes; requiere estructura:leer",
        "responses": {
          "200": {
            "description": "Recurso propio",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Identificador para soporte"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recurso"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida/vencida",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Servicio/operador/plan no habilitado",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no disponible",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto; ver instructivo",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de consultas",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "60 segundos"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible; reintentar sin cambiar IDs",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "registrar_lotes",
        "summary": "Alta repetible de lotes; requiere estructura:crear",
        "description": "No es actualización: otro contenido bajo el mismo ID devuelve 409. No crear IDs nuevos para reintentar un timeout.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alta confirmada o repetida, indicado por creado",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Identificador para soporte"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recurso"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Credencial inválida/vencida",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Servicio/operador/plan no habilitado",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no disponible",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto; ver instructivo",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de consultas",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "60 segundos"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible; reintentar sin cambiar IDs",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "IntegrationKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Clave exclusiva del backend del integrador, nunca contraseña de Chamán."
      }
    },
    "schemas": {
...
      "Servicios": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "version": {
            "type": "string"
          },
          "entorno": {
            "type": "string",
            "enum": ["testing", "production"]
          },
          "integracion": {
            "type": "string"
          },
          "servicios": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "codigo": {
                  "type": "string"
                },
                "operaciones": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": ["codigo", "operaciones"]
            }
          },
          "permisos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "solicitudesPorMinuto": {
            "type": "integer"
          },
          "limites": {
            "type": "object",
            "nullable": true,
            "additionalProperties": false,
            "properties": {
              "productores": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100000
              },
              "establecimientos": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100000
              },
              "lotes": { "type": "integer", "minimum": 0, "maximum": 100000 }
            },
            "required": ["productores", "establecimientos", "lotes"]
          },
          "antiguedadMaximaSiembraDias": {
            "type": "integer",
            "nullable": true,
            "minimum": 1,
            "maximum": 366
          },
          "actualizacion": {
            "type": "string",
            "enum": ["consulta_periodica"]
          },
          "webhooks": {
            "type": "boolean",
            "enum": [false]
          }
        },
        "required": [
          "version",
          "entorno",
          "integracion",
          "servicios",
          "permisos",
          "solicitudesPorMinuto",
          "limites",
          "antiguedadMaximaSiembraDias",
          "actualizacion",
          "webhooks"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}